Modify Categorie.php and CategorieTest.php

+++ app/config/controllers/Categorie.php
@@ -3,24 +3,33 @@
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
+use Appli\PasswordEncoder;
+
 $app->post('/admin/categorie', function (Request $request) use ($app) {
     if (null == $request->getContent()) {
-        return new Response(null, 404);
+        return new Response('No content', 404);
     }
     $categorieArray = json_decode($request->getContent());
+    if (!isset($categorieArray->{'a'})) {
+        return new Response('Admin not connected', 403);
+    }
+    $encoder = new PasswordEncoder();
+    if ($encoder->encodePassword('Admin connected') !== $categorieArray->{'a'}) {
+        return new Response('Admin not connected', 403);
+    }
     if (!isset($categorieArray->{'name_fr'}) || !isset($categorieArray->{'name_en'})) {
-        return new Response(null, 404);
+        return new Response('Attributes name_fr or name_en not here', 404);
     }
     if (!isset($categorieArray->{'ID'})) {
         $sqlRequest = 'INSERT INTO categorie VALUES (null, ?, ?)';
         $app['db']->executeUpdate($sqlRequest, array($categorieArray->{'name_fr'}, $categorieArray->{'name_en'}));
 
-        return new Response(null, 200);
+        return new Response('Categorie created', 200);
     }
     $sqlRequest = 'INSERT INTO categorie VALUES (?, ?, ?)';
     $app['db']->executeUpdate($sqlRequest, array($categorieArray->{'ID'}, $categorieArray->{'name_fr'}, $categorieArray->{'name_en'}));
 
-    return new Response(null, 200);
+    return new Response('Categorie created', 200);
 });
 
 $app->get('/admin/categories', function () use ($app) {
@@ -28,7 +37,7 @@ $app->get('/admin/categories', function () use ($app) {
     $query = $app['db']->executeQuery($sqlRequest);
     $results = $query->fetchAll();
     if (null == $results) {
-        return new Response(null, 400);
+        return new Response('No categories', 400);
     }
     $categories = array();
     foreach ($results as $row) {
@@ -43,7 +52,7 @@ $app->get('/admin/categories/{id}', function ($id) use ($app) {
     $sqlRequest = 'SELECT * FROM categorie WHERE ID = ?';
     $result = $app['db']->fetchAssoc($sqlRequest, array((int) $id));
     if (null == $result) {
-        return new Response(null, 400);
+        return new Response('Categorie don\'t exist', 400);
     }
     $jsonCategorie = json_encode($result);
 
@@ -52,40 +61,69 @@ $app->get('/admin/categories/{id}', function ($id) use ($app) {
 
 $app->put('/admin/categories/{id}', function (Request $request, $id) use ($app) {
     if (null == $request->getContent()) {
-        return new Response(null, 404);
+        return new Response('No content', 404);
+    }
+    $categorieArray = json_decode($request->getContent());
+    if (!isset($categorieArray->{'a'})) {
+        return new Response('Admin not connected', 403);
+    }
+    $encoder = new PasswordEncoder();
+    if ($encoder->encodePassword('Admin connected') !== $categorieArray->{'a'}) {
+        return new Response('Admin not connected', 403);
     }
     $sqlRequest = 'SELECT * FROM categorie WHERE ID = ?';
     $result = $app['db']->fetchAssoc($sqlRequest, array((int) $id));
     if (null == $result) {
-        return new Response(null, 400);
+        return new Response('Categorie don\'t exist', 400);
     }
-    $categorieArray = json_decode($request->getContent());
     if (!isset($categorieArray->{'name_fr'}) || !isset($categorieArray->{'name_en'})) {
-        return new Response(null, 404);
+        return new Response('Attributes name_fr or name_en not here', 404);
     }
     $sqlRequest = 'UPDATE categorie SET name_fr = ?, name_en = ? WHERE ID = ?';
     $app['db']->executeUpdate($sqlRequest, array($categorieArray->{'name_fr'}, $categorieArray->{'name_en'},  $id));
 
-    return new Response(null, 200);
+    return new Response('Categorie updated', 200);
 });
 
-$app->delete('/admin/categories/{id}', function ($id) use ($app) {
+$app->delete('/admin/categories/{id}', function (Request $request, $id) use ($app) {
+    if (null == $request->getContent()) {
+        return new Response('No content', 404);
+    }
+    $array = json_decode($request->getContent());
+    if (!isset($array->{'a'})) {
+        return new Response('Admin not connected', 403);
+    }
+    $encoder = new PasswordEncoder();
+    if ($encoder->encodePassword('Admin connected') !== $array->{'a'}) {
+        return new Response('Admin not connected', 403);
+    }
     $sqlRequest = 'SELECT * FROM categorie WHERE ID = ?';
     $result = $app['db']->fetchAssoc($sqlRequest, array((int) $id));
     if (null == $result) {
-        return new Response(null, 400);
+        return new Response('Categorie don\'t exist', 400);
     }
     $sqlRequest = 'DELETE FROM categorie WHERE ID = ?';
     $app['db']->executeUpdate($sqlRequest, array((int) $id));
 
-    return new Response(null, 200);
+    return new Response('Categorie deleted', 200);
 });
 
-$app->delete('/admin/categories', function () use ($app) {
+$app->delete('/admin/categories', function (Request $request) use ($app) {
+    if (null == $request->getContent()) {
+        return new Response('No content', 404);
+    }
+    $array = json_decode($request->getContent());
+    if (!isset($array->{'a'})) {
+        return new Response('Admin not connected', 403);
+    }
+    $encoder = new PasswordEncoder();
+    if ($encoder->encodePassword('Admin connected') !== $array->{'a'}) {
+        return new Response('Admin not connected', 403);
+    }
     $sqlRequest = 'DELETE FROM categorie WHERE 1';
     $app['db']->executeUpdate($sqlRequest);
 
-    return new Response(null, 200);
+    return new Response('Categories deleted', 200);
 });
 
 return $app;
